# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Attendance Gateway API",
    "version": "1.4.0",
    "description": "Zero-Trust Cryptographic Attendance Gateway.\n\n**Authentication**\n- Most client endpoints accept a shared client API key: header `X-Api-Key: ag_...` (mint from the Admin console → API Keys).\n- Role-protected endpoints (professor/admin) use a JWT: `Authorization: Bearer <token>` from `/api/v1/admin/login` or `/api/v1/auth/login`.\n- Global security: `ApiKey OR bearer`; public operations (health, admin login) override with `[]`.\n\n**Building a client?** See the **Client Integration Guide** (QR / 3-second metronome protocol, biometric Gate 2 ordering, timestamp sealing, 250ms window): https://git.atmyhome.tech/het/attendance-gateway/src/branch/main/docs/CLIENT_INTEGRATION_GUIDE.md\n\n**Error payloads:** every 4xx/5xx returns the envelope `{success:false,error:{code,message,latency_ms?,details?}}`. Full code table + client UI actions: **Error Dictionary** — https://git.atmyhome.tech/het/attendance-gateway/src/branch/main/docs/ERROR_DICTIONARY.md (schema: `ErrorEnvelope`).\n\n**To get an API key:** sign in to the Admin console (https://admin.atmyhome.tech) → **API Keys** → Generate Key."
  },
  "servers": [
    {
      "url": "https://api.atmyhome.tech"
    }
  ],
...
  "security": [
    {
      "ApiKey": []
    },
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Api-Key",
        "description": "Shared client API key (mint from the Admin console)."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/admin/login": {
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Admin login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/admin/login/refresh": {
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Refresh admin token",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/admin/assignments": {
      "get": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "List assignments (timetable)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Create an assignment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prof_uuid": {
                    "type": "string"
                  },
                  "course_code": {
                    "type": "string"
                  },
                  "division_id": {
                    "type": "string"
                  },
                  "day_of_week": {
                    "type": "integer"
                  },
                  "start_time": {
                    "type": "string"
                  },
                  "end_time": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/admin/api-keys": {
      "get": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "List API keys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Generate a new API key",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/api/v1/admin/api-keys/{uuid}/revoke": {
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Revoke an API key",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/admin/change-password": {
      "post": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Change admin password",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "current_password": {
                    "type": "string"
                  },
                  "new_password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/admin/assignments/{id}": {
      "put": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Update an assignment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prof_uuid": {
                    "type": "string"
                  },
                  "course_code": {
                    "type": "string"
                  },
                  "division_id": {
                    "type": "string"
                  },
                  "day_of_week": {
                    "type": "integer"
                  },
                  "start_time": {
                    "type": "string"
                  },
                  "end_time": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Core"
        ],
        "summary": "Delete an assignment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  }
}